refactor: change SelectedCountriesRepository from class to object
